Fixed ASCII simlib output unit of RA, DEC
SNANA SIMLIBS use RA, DEC values in degrees, while this was being
printed out in radians. Changed the code so that the ASCII output is in
radians. Checked that this works correctly by looking at the outut and
comparing to SNANA simlibs.

modified:   ./simlibs/summarize_opsim.py

modified:   WritingOutSimlibs.ipynb
    Nothing really modified