quick and dirty fix for possible channel issue:
I believe this is happening because the conda installs are not
progressing correctly for each channel. The fix I implemented involves
having the conda install script actually run two different install
commands for the different channels. This is also added to
`./setup_generate_requirements.py` which will also have to be changed if
channels are changed. This is sub-optimal and an issue should be created
if this solution is used.
	renamed:    install/conda_requirements.txt -> install/conda_requirements_anaconda.txt
	new file:   install/conda_requirements_openastronomy.txt
	modified:   install/install_conda_requirements.sh
	modified:   install/requirements.md
	modified:   setup/generate_requirements.py

- had a bug and Fixed channel
	modified:   install/spec-file.txt
- Update version
	modified:   opsimsummary/version.py

- update healpy install moving from openastronomy to conda-forge
	renamed:    install/conda_requirements_openastronomy.txt -> install/conda_requirements_conda_forge.txt
	modified:   install/install_conda_requirements.sh